move _describe_drs_object to htsget_operations
