+++ htsget_server/htsget_operations.py
@@ -1,4 +1,5 @@
 import os
+import re
 import tempfile
 from flask import send_file, Flask
 from urllib.parse import urlencode
@@ -149,7 +150,7 @@ def index_analysis(id_=None, force=False, genome='hg38'):
         return {"message": "User is not authorized to index analyses"}, 403
     if id_ is not None:
         # check that there is a database drs object for this:
-        drs_obj = drs_operations._describe_drs_object(id_)
+        drs_obj = _describe_drs_object(id_)
         if drs_obj is None:
             return {"message": f"No DRS object exists with ID {id_}"}, 404
         program = ""
@@ -295,7 +296,7 @@ def _get_experiment(id_=None):
             result["transcriptomes"].append(experiment_drs_obj["id"])
         result["program"] = experiment_drs_obj["program"]
         for contents_obj in experiment_drs_obj["contents"]:
-            drs_obj = drs_operations._describe_drs_object(contents_obj["id"])
+            drs_obj = _describe_drs_object(contents_obj["id"])
             if drs_obj is not None and "type" in drs_obj:
                 if drs_obj["type"] == "variant":
                     result["variants"].append(drs_obj["name"])
@@ -503,7 +504,7 @@ def _get_urls(file_type, id, reference_name=None, start=None, end=None, _class=N
     if file_type not in ["variant", "read"]:
         raise ValueError("File type must be 'variant' or 'read'")
 
-    drs_obj = drs_operations._describe_drs_object(id)
+    drs_obj = _describe_drs_object(id)
     if drs_obj is not None and "status_code" not in drs_obj:
         if "format" not in drs_obj:
             raise Exception(f"no format: {drs_obj}")
@@ -527,7 +528,7 @@ def _get_urls(file_type, id, reference_name=None, start=None, end=None, _class=N
 
 def _verify_analysis_drs_object(id_):
     # get the listed experiments that the AnalysisDrsObject says should be in the file
-    gen_drs_obj = drs_operations._describe_drs_object(id_)
+    gen_drs_obj = _describe_drs_object(id_)
     if gen_drs_obj is None:
         raise Exception(f"Could not find object {id_}")
     drs_experiments = set(gen_drs_obj['experiments'].keys())
@@ -566,7 +567,7 @@ def _verify_analysis_drs_object(id_):
 # two contents objects.
 def get_pysam_obj(object_id):
     result = {'status_code': 200}
-    drs_obj = drs_operations._describe_drs_object(object_id)
+    drs_obj = _describe_drs_object(object_id)
     if drs_obj is None or 'message' in drs_obj:
         return { "message": f"{object_id} not found", "status_code": 404}
     index_result = drs_operations._get_file_path(drs_obj['index'])
@@ -591,4 +592,49 @@ def get_pysam_obj(object_id):
                     result['file'] = VariantFile(main_result['path'], index_filename=index_result['path'])
             except Exception as e:
                 return { "message": str(e), "status_code": 500, "method": f"get_pysam_obj({object_id})"}
+    return result
+
+
+# describe an htsget DRS object, but don't open it
+def _describe_drs_object(object_id):
+    drs_obj = drs_database.get_drs_object(object_id)
+    if drs_obj is None:
+        return None
+    result = {
+        "name": object_id,
+        "program": drs_obj["program"]
+    }
+    # drs_obj should have a main contents, index contents, and experiment contents
+    if "contents" in drs_obj:
+        for contents in drs_obj["contents"]:
+            # get each drs object (should be the analysis file and its index)
+            # if sub_obj.name matches an index file regex, it's an index file
+            index_match = re.fullmatch(r'.+\.(...*i)$', contents["name"])
+
+            # if sub_obj.name matches a bam/sam/cram file regex, it's a read file
+            read_match = re.fullmatch(r'.+\.(.+?am)$', contents["name"])
+
+            # if sub_obj.name matches a vcf/bcf file regex, it's a variant file
+            variant_match = re.fullmatch(r'.+\.(.cf)(\.gz)*$', contents["name"])
+
+            if read_match is not None:
+                result['format'] = read_match.group(1).upper()
+                result['type'] = "read"
+                result['main'] = contents['name']
+            elif variant_match is not None:
+                result['format'] = variant_match.group(1).upper()
+                result['type'] = "variant"
+                result['main'] = contents['name']
+            elif index_match is not None:
+                result['index'] = contents['name']
+            else:
+                ## this is for migration: experiments used to be samples
+                if "samples" in result:
+                    result["experiments"] = result["samples"]
+                if "experiments" not in result:
+                    result['experiments'] = {}
+                result['experiments'][contents['id']] = contents['name']
+
+    if 'type' not in result:
+        return {"message": f"drs object {object_id} does not represent an htsget object", "status_code": 404}
     return result
